+++ packages/rspeedy/core/src/cli/build.ts
@@ -6,12 +6,15 @@ import { logger } from '@rsbuild/core'
 import type { Command } from 'commander'
 
 import type { CommonOptions } from './commands.js'
+import type { CreateRspeedyOptions } from '../create-rspeedy.js'
 import { exit } from './exit.js'
 import { loadConfig } from '../config/loadConfig.js'
 import { createRspeedy } from '../create-rspeedy.js'
 import { isCI } from '../utils/is-ci.js'
 
-export type BuildOptions = CommonOptions
+export type BuildOptions = CommonOptions & {
+  environment?: string[] | undefined
+}
 
 export async function build(
   this: Command,
@@ -28,7 +31,20 @@ export async function build(
       configPath: buildOptions.config,
     })
 
-    const rspeedy = await createRspeedy({ cwd, rspeedyConfig })
+    const options: CreateRspeedyOptions = {
+      cwd,
+      rspeedyConfig,
+    }
+
+    if (buildOptions.envMode) {
+      options.loadEnv = { mode: buildOptions.envMode }
+    }
+
+    if (buildOptions.environment) {
+      options.environment = buildOptions.environment
+    }
+
+    const rspeedy = await createRspeedy(options)
 
     await rspeedy.build()
   } catch (error) {