+++ packages/rspeedy/core/src/plugins/chunkLoading.plugin.ts
@@ -6,31 +6,34 @@ import type { RsbuildPlugin } from '@rsbuild/core'
 
 import { ChunkLoadingWebpackPlugin } from '@lynx-js/chunk-loading-webpack-plugin'
 
+import { isLynx } from '../utils/is-lynx.js'
+import { isWeb } from '../utils/is-web.js'
+
 export function pluginChunkLoading(): RsbuildPlugin {
   return {
     name: 'lynx:rsbuild:chunk-loading',
     setup(api) {
-      api.modifyBundlerChain(chain => {
-        // dprint-ignore
-        chain
+      api.modifyBundlerChain((chain, { environment }) => {
+        if (isWeb(environment)) {
+          chain
+            .output
+            .chunkLoading('import-scripts')
+            .end()
+          return
+        }
+
+        if (isLynx(environment)) {
+          // dprint-ignore
+          chain
           .plugin('lynx:chunk-loading')
             .use(ChunkLoadingWebpackPlugin)
           .end()
           .output
-            // Rspack needs `chunkLoading: 'require'` since we use runtimeModule hook
-            // to override the chunk loading runtime.
-            .chunkLoading('require')
+            .chunkLoading('lynx')
             .chunkFormat('commonjs')
             .iife(false)
           .end()
-      })
-
-      api.modifyWebpackChain(chain => {
-        chain
-          .output
-          // For webpack, we directly use `chunkLoading: 'lynx'`.
-          .chunkLoading('lynx')
-          .end()
+        }
       })
     },
   }