+++ packages/webpack/css-extract-webpack-plugin/test/hotCases/hot-update-json/default/rspack.config.js
@@ -4,13 +4,14 @@
 // LICENSE file in the root directory of this source tree.
 */
 import rspack from '@rspack/core'
-import { CssExtractRspackPlugin } from '../../../../src/index.js'
+import { CssExtractRspackPlugin } from '../../../../src/index'
 import path from 'node:path'
 
 /** @type {import('webpack').Configuration} */
 export default {
   output: {
     publicPath: './',
+    pathinfo: false,
   },
   module: {
     parser: {