+++ packages/rspeedy/plugin-react-alias/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @lynx-js/react-alias-rsbuild-plugin
 
+## 0.9.5
+
+## 0.9.4
+
 ## 0.9.3
 
 ### Patch Changes