+++ packages/rspeedy/plugin-react/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @lynx-js/react-rsbuild-plugin
 
+## 0.9.5
+
+### Patch Changes
+
+- fix: add enableCSSInvalidation for encodeCSS of css HMR, this will fix pseudo-class (such as `:active`) not working in HMR. ([#435](https://github.com/lynx-family/lynx-stack/pull/435))
+
+- Disable `module.generator.json.JSONParse` option as it increases the bundle size of `main-thread.js`. For more detail, please see this [issue](https://github.com/webpack/webpack/issues/19319). ([#402](https://github.com/lynx-family/lynx-stack/pull/402))
+
+- Updated dependencies [[`3e7988f`](https://github.com/lynx-family/lynx-stack/commit/3e7988f3af4b4f460eaf5add29cca19537dc1a6b), [`7243242`](https://github.com/lynx-family/lynx-stack/commit/7243242801e3a8ca0213c0ef642f69a22c39960e)]:
+  - @lynx-js/css-extract-webpack-plugin@0.5.3
+  - @lynx-js/template-webpack-plugin@0.6.8
+  - @lynx-js/react-alias-rsbuild-plugin@0.9.5
+  - @lynx-js/react-refresh-webpack-plugin@0.3.2
+  - @lynx-js/react-webpack-plugin@0.6.10
+  - @lynx-js/web-webpack-plugin@0.6.3
+
+## 0.9.4
+
+### Patch Changes
+
+- feat: add extractStr option to pluginReactLynx ([#391](https://github.com/lynx-family/lynx-stack/pull/391))
+
+- Convert background-only files from js to ts ([#346](https://github.com/lynx-family/lynx-stack/pull/346))
+
+- Updated dependencies [[`f849117`](https://github.com/lynx-family/lynx-stack/commit/f84911731faa4d0f6373d1202b9b2cabb0bafc48), [`d730101`](https://github.com/lynx-family/lynx-stack/commit/d7301017a383b8825cdc813a649ef26ce1c37641), [`42217c2`](https://github.com/lynx-family/lynx-stack/commit/42217c2c77a33e729977fc7108b218a1cb868e6a), [`f03bd4a`](https://github.com/lynx-family/lynx-stack/commit/f03bd4a62f81902ba55caf10df56447c89743e62)]:
+  - @lynx-js/react-webpack-plugin@0.6.10
+  - @lynx-js/template-webpack-plugin@0.6.7
+  - @lynx-js/react-alias-rsbuild-plugin@0.9.4
+  - @lynx-js/react-refresh-webpack-plugin@0.3.2
+  - @lynx-js/css-extract-webpack-plugin@0.5.2
+  - @lynx-js/web-webpack-plugin@0.6.3
+
 ## 0.9.3
 
 ### Patch Changes