+++ packages/web-platform/web-elements/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @lynx-js/web-elements
 
+## 0.5.3
+
+### Patch Changes
+
+- feat: add `layoutchange` event support for x-view and x-text ([#408](https://github.com/lynx-family/lynx-stack/pull/408))
+
+## 0.5.2
+
+### Patch Changes
+
+- fix: When list with list-type: flow, scrolltoupper, scrolltolower were specified, there was a blank block. ([#379](https://github.com/lynx-family/lynx-stack/pull/379))
+
+- fix: do not show scroll bar ([#406](https://github.com/lynx-family/lynx-stack/pull/406))
+
 ## 0.5.1
 
 ### Patch Changes